Organize pieces into own package
