+++ .gitignore
@@ -46,5 +46,8 @@ Network Trash Folder
 Temporary Items
 .apdisk
 
+bin/
 out/
 .idea/
+.classpath
+.project
