+++ src/com/hbh7/pieces/Queen.java
@@ -1,28 +1,29 @@
-package com.hbh7;
+package com.hbh7.pieces;
 
 import static com.hbh7.Util.*;
 
-public class Rook extends ChessPiece{
+public class Queen extends ChessPiece{
 
-    private String pieceType = "Rook";
+    private String pieceType = "Queen";
     private String owner;
 
-    public Rook(String owner, String position) {
-        super("Rook", owner, position);
+    public Queen(String owner, String position) {
+        super("Queen", owner, position);
         this.owner = owner;
-        this.pieceValue = 4;
+        this.pieceValue = 6;
     }
 
     public boolean checkValidMove_movePatternValidCheck(int playerNum, int originalRow, String originalColumn, int newRow, String newColumn, ChessPiece[][] boardArray) {
 
         // Check that the move pattern requested is valid
-        // Ex: Rook can move forward, backwards, left, or right any number of spaces until it collides with the board
-        // border or another piece.
+        // Ex: Queen can move in any direction forwards, backwards, left, right, or diagonal an unlimited number of spaces
+        // until it hits a border or another piece. It cannot move like a Knight.
 
+
+        // Rook bit
 
         if((originalColumn.equals(newColumn)) ^ (originalRow == newRow)) { // bitwise XOR, must change only one to be true
             if(originalColumn.equals(newColumn)) {
-                System.out.println("Row is changing, piece is moving vertical");
                 // Row is changing, piece is moving vertical
                 if (toArrayIndex(originalRow) < toArrayIndex(newRow)) { // If moving downwards
                     boolean error = false;
@@ -32,7 +33,7 @@ public class Rook extends ChessPiece{
                         }
                     }
                     if (error) {
-                        System.out.println("Rook: Invalid Move. Piece(s) obstructing path.");
+                        System.out.println("Queen: Invalid Move. Piece(s) obstructing path.");
                         return false;
                     } else {
                         return true;
@@ -45,7 +46,7 @@ public class Rook extends ChessPiece{
                         }
                     }
                     if (error) {
-                        System.out.println("Rook: Invalid Move. Piece(s) obstructing path.");
+                        System.out.println("Queen: Invalid Move. Piece(s) obstructing path.");
                         return false;
                     } else {
                         return true;
@@ -62,7 +63,7 @@ public class Rook extends ChessPiece{
                         }
                     }
                     if (error) {
-                        System.out.println("Rook: Invalid Move. Piece(s) obstructing path.");
+                        System.out.println("Queen: Invalid Move. Piece(s) obstructing path.");
                         return false;
                     } else {
                         return true;
@@ -75,18 +76,65 @@ public class Rook extends ChessPiece{
                         }
                     }
                     if (error) {
-                        System.out.println("Rook: Invalid Move. Piece(s) obstructing path.");
+                        System.out.println("Queen: Invalid Move. Piece(s) obstructing path.");
                         return false;
                     } else {
                         return true;
                     }
                 }
             }
+        } else if(Math.abs(toArrayIndex(originalRow)-toArrayIndex(newRow)) == Math.abs(toArrayIndex(originalColumn)-toArrayIndex(newColumn))) { // Bishop Bit
+
+            int distance = Math.abs(toArrayIndex(originalRow) - toArrayIndex(newRow));
+
+            if (toArrayIndex(newRow) < toArrayIndex(originalRow) && toArrayIndex(newColumn) > toArrayIndex(originalColumn)) { // going to top right
+                for (int i = 1; i < distance; i++) {
+                    if (boardArray[toArrayIndex(originalRow) - i][toArrayIndex(originalColumn) + i] != null) {
+                        System.out.println("Queen: Invalid Move. Piece(s) obstructing path.");
+                        return false;
+                    }
+                }
+                return true;
+
+            } else if (toArrayIndex(newRow) < toArrayIndex(originalRow) && toArrayIndex(newColumn) < toArrayIndex(originalColumn)) { // going to top left
+                for (int i = 1; i < distance; i++) {
+                    if (boardArray[toArrayIndex(originalRow) - i][toArrayIndex(originalColumn) - i] != null) {
+                        System.out.println("Queen: Invalid Move. Piece(s) obstructing path.");
+                        return false;
+                    }
+                }
+                return true;
+
+            } else if (toArrayIndex(newRow) > toArrayIndex(originalRow) && toArrayIndex(newColumn) > toArrayIndex(originalColumn)) { // going to bottom right
+                for (int i = 1; i < distance; i++) {
+                    if (boardArray[toArrayIndex(originalRow) + i][toArrayIndex(originalColumn) + i] != null) {
+                        System.out.println("Queen: Invalid Move. Piece(s) obstructing path.");
+                        return false;
+                    }
+                }
+                return true;
+
+            } else if (toArrayIndex(newRow) > toArrayIndex(originalRow) && toArrayIndex(newColumn) < toArrayIndex(originalColumn)) { // going to bottom left
+                for (int i = 1; i < distance; i++) {
+                    if (boardArray[toArrayIndex(originalRow) + i][toArrayIndex(originalColumn) - i] != null) {
+                        System.out.println("Queen: Invalid Move. Piece(s) obstructing path.");
+                        return false;
+                    }
+                }
+                return true;
+
+            } else {
+                System.out.println("Queen: Honestly I have no idea how you got here.");
+                return false;
+            }
         } else {
-            System.out.println("Rook: Invalid Move. Cannot move across 2 axis at once.");
             return false;
         }
+
+
+
     }
+
 
     public PieceData aiFindSpacesToAttack(ChessPiece[][] boardArray) {
         // Pawn code, revise for specific moves
@@ -104,4 +152,5 @@ public class Rook extends ChessPiece{
         }
         return null;
     }
+
 }
